+++ tests/object/test_upload.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: GPL-2.0
 
 from uhu.core.object import Object
-from uhu.core.upload import ObjectUploadResult
+from uhu.core.updatehub import ObjectUploadResult
 from uhu.exceptions import UploadError
 from uhu.utils import CHUNK_SIZE_VAR, SERVER_URL_VAR
